Vakken bij groepen create, edit Viewmodel bij vakken, met bijbehorende studenten (overzicht) jquery vailidator bij resultaat invoegen, lijst met studenten met bijbehorende resultaten bij resultaten view. kleine aanpassingen aan koppeltabel views, en stijling.

+++ ResultatenSysteem/ResultatenSysteem/ResultatenSysteem/Models/Resultaat.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace ResultatenSysteem.Models
     public class Resultaat
     {
         public int Id { get; set; }
-        public int Beoordeling { get; set; }
+        public double Beoordeling { get; set; }
 
         public Student Student { get; set; }
         public int StudentId { get; set; }
